move wireCompat to hosting
