+++ packages/agents-activity/src/index.ts
@@ -42,4 +42,3 @@ export { MessageReaction } from './messageReaction'
 export { MessageReactionTypes } from './messageReactionTypes'
 export { TextFormatTypes } from './textFormatTypes'
 export { TextHighlight } from './textHighlight'
-export * from './activityWireCompat'
