+++ packages/agents-hosting/src/activityWireCompat.ts
